wrapper function to call limiter computation, compilation in only one cpp
