+++ SU2_CFD/include/variables/CEulerVariable.hpp
@@ -115,6 +115,12 @@ public:
   }
 
   /*!
+   * \brief Get the primitive variables limiter.
+   * \return Primitive variables limiter for the entire domain.
+   */
+  inline MatrixType* GetLimiter_Primitive(void) {return &Limiter_Primitive; }
+
+  /*!
    * \brief Get the value of the primitive variables gradient.
    * \param[in] iVar - Index of the variable.
    * \return Value of the primitive variables gradient.
