+++ SU2_CFD/src/solver_direct_mean_inc.cpp
@@ -32,6 +32,7 @@
 #include "../include/variables/CIncNSVariable.hpp"
 #include "../include/gradients/computeGradientsGreenGauss.hpp"
 #include "../include/gradients/computeGradientsLeastSquares.hpp"
+#include "../include/limiters/computeLimiters.hpp"
 
 
 CIncEulerSolver::CIncEulerSolver(void) : CSolver() {
@@ -3594,299 +3595,16 @@ void CIncEulerSolver::SetPrimitive_Gradient_LS(CGeometry *geometry, CConfig *con
 }
 
 void CIncEulerSolver::SetPrimitive_Limiter(CGeometry *geometry, CConfig *config) {
-  
-  unsigned long iEdge, iPoint, jPoint;
-  unsigned short iVar, iDim;
-  su2double **Gradient_i, **Gradient_j, *Coord_i, *Coord_j,
-  *Primitive, *Primitive_i, *Primitive_j,
-  *LocalMinPrimitive = NULL, *LocalMaxPrimitive = NULL,
-  *GlobalMinPrimitive = NULL, *GlobalMaxPrimitive = NULL,
-  dave, LimK, eps2, eps1, dm, dp, du, y, limiter;
-  
-#ifdef CODI_REVERSE_TYPE
-  bool TapeActive = false;
 
-  if (config->GetDiscrete_Adjoint() && config->GetFrozen_Limiter_Disc()) {
-    /*--- If limiters are frozen do not record the computation ---*/
-    TapeActive = AD::globalTape.isActive();
-    AD::StopRecording();
-  }
-#endif
-  
-  dave = config->GetRefElemLength();
-  LimK = config->GetVenkat_LimiterCoeff();
+  auto kindLimiter = static_cast<ENUM_LIMITER>(config->GetKind_SlopeLimit_Flow());
+  auto primitives = nodes->GetPrimitive();
+  auto gradient = nodes->GetGradient_Reconstruction();
+  auto primMin = nodes->GetSolution_Min();
+  auto primMax = nodes->GetSolution_Max();
+  auto limiter = nodes->GetLimiter_Primitive();
 
-  if (config->GetKind_SlopeLimit_Flow() == NO_LIMITER) {
-   
-    for (iPoint = 0; iPoint < geometry->GetnPoint(); iPoint++) {
-      for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-        nodes->SetLimiter_Primitive(iPoint,iVar, 1.0);
-      }
-    }
-    
-  }
-  
-  else {
-    
-    /*--- Initialize solution max and solution min and the limiter in the entire domain --*/
-    
-    for (iPoint = 0; iPoint < geometry->GetnPoint(); iPoint++) {
-      for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-        nodes->SetSolution_Max(iPoint,iVar, -EPS);
-        nodes->SetSolution_Min(iPoint,iVar, EPS);
-        nodes->SetLimiter_Primitive(iPoint,iVar, 2.0);
-      }
-    }
-    
-    /*--- Establish bounds for Spekreijse monotonicity by finding max & min values of neighbor variables --*/
-    
-    for (iEdge = 0; iEdge < geometry->GetnEdge(); iEdge++) {
-      
-      /*--- Point identification, Normal vector and area ---*/
-      
-      iPoint = geometry->edge[iEdge]->GetNode(0);
-      jPoint = geometry->edge[iEdge]->GetNode(1);
-      
-      /*--- Get the primitive variables ---*/
-      
-      Primitive_i = nodes->GetPrimitive(iPoint);
-      Primitive_j = nodes->GetPrimitive(jPoint);
-      
-      /*--- Compute the maximum, and minimum values for nodes i & j ---*/
-      
-      for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-        du = (Primitive_j[iVar] - Primitive_i[iVar]);
-        nodes->SetSolution_Min(iPoint, iVar, min(nodes->GetSolution_Min(iPoint, iVar), du));
-        nodes->SetSolution_Max(iPoint, iVar, max(nodes->GetSolution_Max(iPoint, iVar), du));
-        nodes->SetSolution_Min(jPoint, iVar, min(nodes->GetSolution_Min(jPoint, iVar), -du));
-        nodes->SetSolution_Max(jPoint, iVar, max(nodes->GetSolution_Max(jPoint, iVar), -du));
-      }
-      
-    }
-    
-    /*--- Correct the limiter values across any periodic boundaries. ---*/
-
-    for (unsigned short iPeriodic = 1; iPeriodic <= config->GetnMarker_Periodic()/2; iPeriodic++) {
-      InitiatePeriodicComms(geometry, config, iPeriodic, PERIODIC_LIM_PRIM_1);
-      CompletePeriodicComms(geometry, config, iPeriodic, PERIODIC_LIM_PRIM_1);
-    }
-    
-  }
-  
-  
-  /*--- Barth-Jespersen limiter with Venkatakrishnan modification ---*/
-  
-  if (config->GetKind_SlopeLimit_Flow() == BARTH_JESPERSEN) {
-    
-    for (iEdge = 0; iEdge < geometry->GetnEdge(); iEdge++) {
-      
-      iPoint     = geometry->edge[iEdge]->GetNode(0);
-      jPoint     = geometry->edge[iEdge]->GetNode(1);
-      Gradient_i = nodes->GetGradient_Reconstruction(iPoint);
-      Gradient_j = nodes->GetGradient_Reconstruction(jPoint);
-      Coord_i    = geometry->node[iPoint]->GetCoord();
-      Coord_j    = geometry->node[jPoint]->GetCoord();
-      
-      AD::StartPreacc();
-      AD::SetPreaccIn(Gradient_i, nPrimVarGrad, nDim);
-      AD::SetPreaccIn(Gradient_j, nPrimVarGrad, nDim);
-      AD::SetPreaccIn(Coord_i, nDim); AD::SetPreaccIn(Coord_j, nDim);
-
-      for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-        
-        AD::SetPreaccIn(nodes->GetSolution_Max(iPoint,iVar));
-        AD::SetPreaccIn(nodes->GetSolution_Min(iPoint,iVar));
-        AD::SetPreaccIn(nodes->GetSolution_Max(jPoint,iVar));
-        AD::SetPreaccIn(nodes->GetSolution_Min(jPoint,iVar));
-
-        /*--- Calculate the interface left gradient, delta- (dm) ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_j[iDim]-Coord_i[iDim])*Gradient_i[iVar][iDim];
-        
-        if (dm == 0.0) { limiter = 2.0; }
-        else {
-          if ( dm > 0.0 ) dp = nodes->GetSolution_Max(iPoint,iVar);
-          else dp = nodes->GetSolution_Min(iPoint,iVar);
-          limiter = dp/dm;
-        }
-        
-        if (limiter < nodes->GetLimiter_Primitive(iPoint,iVar)) {
-          nodes->SetLimiter_Primitive(iPoint,iVar, limiter);
-          AD::SetPreaccOut(nodes->GetLimiter_Primitive(iPoint)[iVar]);
-        }
-        
-        /*--- Calculate the interface right gradient, delta+ (dp) ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_i[iDim]-Coord_j[iDim])*Gradient_j[iVar][iDim];
-        
-        if (dm == 0.0) { limiter = 2.0; }
-        else {
-          if ( dm > 0.0 ) dp = nodes->GetSolution_Max(jPoint,iVar);
-          else dp = nodes->GetSolution_Min(jPoint,iVar);
-          limiter = dp/dm;
-        }
-        
-        if (limiter < nodes->GetLimiter_Primitive(jPoint,iVar)) {
-          nodes->SetLimiter_Primitive(jPoint,iVar, limiter);
-          AD::SetPreaccOut(nodes->GetLimiter_Primitive(jPoint)[iVar]);
-        }
-        
-      }
-      
-      AD::EndPreacc();
-      
-    }
-    
-    for (iPoint = 0; iPoint < geometry->GetnPoint(); iPoint++) {
-      for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-        y =  nodes->GetLimiter_Primitive(iPoint,iVar);
-        limiter = (y*y + 2.0*y) / (y*y + y + 2.0);
-        nodes->SetLimiter_Primitive(iPoint,iVar, limiter);
-      }
-    }
-    
-  }
-  
-  /*--- Venkatakrishnan limiter ---*/
-  
-  if ((config->GetKind_SlopeLimit_Flow() == VENKATAKRISHNAN) ||
-      (config->GetKind_SlopeLimit_Flow() == VENKATAKRISHNAN_WANG)) {
-    
-    if (config->GetKind_SlopeLimit_Flow() == VENKATAKRISHNAN_WANG) {
-      
-      /*--- Allocate memory for the max and min primitive value --*/
-      
-      LocalMinPrimitive = new su2double [nPrimVarGrad]; GlobalMinPrimitive = new su2double [nPrimVarGrad];
-      LocalMaxPrimitive = new su2double [nPrimVarGrad]; GlobalMaxPrimitive = new su2double [nPrimVarGrad];
-      
-      /*--- Compute the max value and min value of the solution ---*/
-      
-      Primitive = nodes->GetPrimitive(0);
-      for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-        LocalMinPrimitive[iVar] = Primitive[iVar];
-        LocalMaxPrimitive[iVar] = Primitive[iVar];
-      }
-      
-      for (iPoint = 0; iPoint < geometry->GetnPoint(); iPoint++) {
-        
-        /*--- Get the primitive variables ---*/
-        
-        Primitive = nodes->GetPrimitive(iPoint);
-
-        for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-          LocalMinPrimitive[iVar] = min (LocalMinPrimitive[iVar], Primitive[iVar]);
-          LocalMaxPrimitive[iVar] = max (LocalMaxPrimitive[iVar], Primitive[iVar]);
-        }
-        
-      }
-
-#ifdef HAVE_MPI
-      SU2_MPI::Allreduce(LocalMinPrimitive, GlobalMinPrimitive, nPrimVarGrad, MPI_DOUBLE, MPI_MIN, MPI_COMM_WORLD);
-      SU2_MPI::Allreduce(LocalMaxPrimitive, GlobalMaxPrimitive, nPrimVarGrad, MPI_DOUBLE, MPI_MAX, MPI_COMM_WORLD);
-#else
-      for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-        GlobalMinPrimitive[iVar] = LocalMinPrimitive[iVar];
-        GlobalMaxPrimitive[iVar] = LocalMaxPrimitive[iVar];
-      }
-#endif
-    }
-    
-    for (iEdge = 0; iEdge < geometry->GetnEdge(); iEdge++) {
-      
-      iPoint     = geometry->edge[iEdge]->GetNode(0);
-      jPoint     = geometry->edge[iEdge]->GetNode(1);
-      Gradient_i = nodes->GetGradient_Primitive(iPoint);
-      Gradient_j = nodes->GetGradient_Primitive(jPoint);
-      Coord_i    = geometry->node[iPoint]->GetCoord();
-      Coord_j    = geometry->node[jPoint]->GetCoord();
-      
-      for (iVar = 0; iVar < nPrimVarGrad; iVar++) {
-          
-        AD::StartPreacc();
-        AD::SetPreaccIn(Gradient_i[iVar], nDim);
-        AD::SetPreaccIn(Gradient_j[iVar], nDim);
-        AD::SetPreaccIn(Coord_i, nDim);
-        AD::SetPreaccIn(Coord_j, nDim);
-        AD::SetPreaccIn(nodes->GetSolution_Max(iPoint,iVar));
-        AD::SetPreaccIn(nodes->GetSolution_Min(iPoint,iVar));
-        AD::SetPreaccIn(nodes->GetSolution_Max(jPoint,iVar));
-        AD::SetPreaccIn(nodes->GetSolution_Min(jPoint,iVar));
-        
-        if (config->GetKind_SlopeLimit_Flow() == VENKATAKRISHNAN_WANG) {
-          AD::SetPreaccIn(GlobalMaxPrimitive[iVar]);
-          AD::SetPreaccIn(GlobalMinPrimitive[iVar]);
-          eps1 = LimK * (GlobalMaxPrimitive[iVar] - GlobalMinPrimitive[iVar]);
-          eps2 = eps1*eps1;
-        }
-        else {
-          eps1 = LimK*dave;
-          eps2 = eps1*eps1*eps1;
-        }
-        
-        /*--- Calculate the interface left gradient, delta- (dm) ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_j[iDim]-Coord_i[iDim])*Gradient_i[iVar][iDim];
-        
-        /*--- Calculate the interface right gradient, delta+ (dp) ---*/
-        
-        if ( dm > 0.0 ) dp = nodes->GetSolution_Max(iPoint,iVar);
-        else dp = nodes->GetSolution_Min(iPoint,iVar);
-        
-        limiter = ( dp*dp + 2.0*dp*dm + eps2 )/( dp*dp + dp*dm + 2.0*dm*dm + eps2);
-        
-        if (limiter < nodes->GetLimiter_Primitive(iPoint,iVar)){
-          nodes->SetLimiter_Primitive(iPoint,iVar, limiter);
-          AD::SetPreaccOut(nodes->GetLimiter_Primitive(iPoint)[iVar]);
-        }
-        
-        /*-- Repeat for point j on the edge ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_i[iDim]-Coord_j[iDim])*Gradient_j[iVar][iDim];
-        
-        if ( dm > 0.0 ) dp = nodes->GetSolution_Max(jPoint,iVar);
-        else dp = nodes->GetSolution_Min(jPoint,iVar);
-        
-        limiter = ( dp*dp + 2.0*dp*dm + eps2 )/( dp*dp + dp*dm + 2.0*dm*dm + eps2);
-        
-        if (limiter < nodes->GetLimiter_Primitive(jPoint,iVar)){
-          nodes->SetLimiter_Primitive(jPoint,iVar, limiter);
-          AD::SetPreaccOut(nodes->GetLimiter_Primitive(jPoint)[iVar]);
-        }
-        
-        AD::EndPreacc();
-      }
-    }
-    
-    if (LocalMinPrimitive  != NULL) delete [] LocalMinPrimitive;
-    if (LocalMaxPrimitive  != NULL) delete [] LocalMaxPrimitive;
-    if (GlobalMinPrimitive != NULL) delete [] GlobalMinPrimitive;
-    if (GlobalMaxPrimitive != NULL) delete [] GlobalMaxPrimitive;
-
-  }
-  
-  /*--- Correct the limiter values across any periodic boundaries. ---*/
-
-  for (unsigned short iPeriodic = 1; iPeriodic <= config->GetnMarker_Periodic()/2; iPeriodic++) {
-    InitiatePeriodicComms(geometry, config, iPeriodic, PERIODIC_LIM_PRIM_2);
-    CompletePeriodicComms(geometry, config, iPeriodic, PERIODIC_LIM_PRIM_2);
-  }
-  
-  /*--- Limiter MPI ---*/
-  
-  InitiateComms(geometry, config, PRIMITIVE_LIMITER);
-  CompleteComms(geometry, config, PRIMITIVE_LIMITER);
-
-#ifdef CODI_REVERSE_TYPE
-  if (TapeActive) AD::StartRecording();
-#endif
+  computeLimiters(kindLimiter, this, PRIMITIVE_LIMITER, PERIODIC_LIM_PRIM_1, PERIODIC_LIM_PRIM_2,
+    *geometry, *config, 0, nPrimVarGrad, *primitives, *gradient, *primMin, *primMax, *limiter);
 }
 
 void CIncEulerSolver::SetInletAtVertex(su2double *val_inlet,
