+++ SU2_CFD/src/solver_structure.cpp
@@ -30,6 +30,7 @@
 #include "../include/variables/CBaselineVariable.hpp"
 #include "../include/gradients/computeGradientsGreenGauss.hpp"
 #include "../include/gradients/computeGradientsLeastSquares.hpp"
+#include "../include/limiters/computeLimiters.hpp"
 #include "../../Common/include/toolboxes/MMS/CIncTGVSolution.hpp"
 #include "../../Common/include/toolboxes/MMS/CInviscidVortexSolution.hpp"
 #include "../../Common/include/toolboxes/MMS/CMMSIncEulerSolution.hpp"
@@ -3072,447 +3073,16 @@ void CSolver::SetAuxVar_Surface_Gradient(CGeometry *geometry, CConfig *config) {
 }
 
 void CSolver::SetSolution_Limiter(CGeometry *geometry, CConfig *config) {
-  
-  unsigned long iEdge, iPoint, jPoint;
-  unsigned short iVar, iDim;
-  su2double **Gradient_i, **Gradient_j, *Coord_i, *Coord_j,
-  *Solution, *Solution_i, *Solution_j,
-  *LocalMinSolution = NULL, *LocalMaxSolution = NULL,
-  *GlobalMinSolution = NULL, *GlobalMaxSolution = NULL,
-  dave, LimK, eps1, eps2, dm, dp, du, ds, y, limiter, SharpEdge_Distance;
-  
-#ifdef CODI_REVERSE_TYPE
-  bool TapeActive = false;
 
-  if (config->GetDiscrete_Adjoint() && config->GetFrozen_Limiter_Disc()) {
-    /*--- If limiters are frozen do not record the computation ---*/
-    TapeActive = AD::globalTape.isActive();
-    AD::StopRecording();
-  }
-#endif
-  
-  dave = config->GetRefElemLength();
-  LimK = config->GetVenkat_LimiterCoeff();
-  
-  if (config->GetKind_SlopeLimit() == NO_LIMITER) {
-    
-    for (iPoint = 0; iPoint < geometry->GetnPoint(); iPoint++) {
-      for (iVar = 0; iVar < nVar; iVar++) {
-        base_nodes->SetLimiter(iPoint, iVar, 1.0);
-      }
-    }
-    
-  }
-  
-  else {
-    
-    /*--- Initialize solution max and solution min and the limiter in the entire domain --*/
-    
-    for (iPoint = 0; iPoint < geometry->GetnPoint(); iPoint++) {
-      for (iVar = 0; iVar < nVar; iVar++) {
-        base_nodes->SetSolution_Max(iPoint, iVar, -EPS);
-        base_nodes->SetSolution_Min(iPoint, iVar, EPS);
-        base_nodes->SetLimiter(iPoint, iVar, 2.0);
-      }
-    }
-    
-    /*--- Establish bounds for Spekreijse monotonicity by finding max & min values of neighbor variables --*/
-    
-    for (iEdge = 0; iEdge < geometry->GetnEdge(); iEdge++) {
-      
-      /*--- Point identification, Normal vector and area ---*/
-      
-      iPoint = geometry->edge[iEdge]->GetNode(0);
-      jPoint = geometry->edge[iEdge]->GetNode(1);
-      
-      /*--- Get the conserved variables ---*/
-      
-      Solution_i = base_nodes->GetSolution(iPoint);
-      Solution_j = base_nodes->GetSolution(jPoint);
-      
-      /*--- Compute the maximum, and minimum values for nodes i & j ---*/
-      
-      for (iVar = 0; iVar < nVar; iVar++) {
-        du = (Solution_j[iVar] - Solution_i[iVar]);
-        base_nodes->SetSolution_Min(iPoint, iVar, min(base_nodes->GetSolution_Min(iPoint, iVar), du));
-        base_nodes->SetSolution_Max(iPoint, iVar, max(base_nodes->GetSolution_Max(iPoint, iVar), du));
-        base_nodes->SetSolution_Min(jPoint, iVar, min(base_nodes->GetSolution_Min(jPoint, iVar), -du));
-        base_nodes->SetSolution_Max(jPoint, iVar, max(base_nodes->GetSolution_Max(jPoint, iVar), -du));
-      }
-      
-    }
-    
-    /*--- Correct the limiter values across any periodic boundaries. ---*/
-    
-    for (unsigned short iPeriodic = 1; iPeriodic <= config->GetnMarker_Periodic()/2; iPeriodic++) {
-      InitiatePeriodicComms(geometry, config, iPeriodic, PERIODIC_LIM_SOL_1);
-      CompletePeriodicComms(geometry, config, iPeriodic, PERIODIC_LIM_SOL_1);
-    }
-    
-  }
-  
-  /*--- Barth-Jespersen limiter with Venkatakrishnan modification ---*/
-  
-  if (config->GetKind_SlopeLimit_Flow() == BARTH_JESPERSEN) {
-    
-    for (iEdge = 0; iEdge < geometry->GetnEdge(); iEdge++) {
-      
-      iPoint     = geometry->edge[iEdge]->GetNode(0);
-      jPoint     = geometry->edge[iEdge]->GetNode(1);
-      Gradient_i = base_nodes->GetGradient_Reconstruction(iPoint);
-      Gradient_j = base_nodes->GetGradient_Reconstruction(jPoint);
-      Coord_i    = geometry->node[iPoint]->GetCoord();
-      Coord_j    = geometry->node[jPoint]->GetCoord();
-      
-      AD::StartPreacc();
-      AD::SetPreaccIn(Gradient_i, nVar, nDim);
-      AD::SetPreaccIn(Gradient_j, nVar, nDim);
-      AD::SetPreaccIn(Coord_i, nDim); AD::SetPreaccIn(Coord_j, nDim);
+  auto kindLimiter = static_cast<ENUM_LIMITER>(config->GetKind_SlopeLimit());
+  auto solution = base_nodes->GetSolution();
+  auto gradient = base_nodes->GetGradient_Reconstruction();
+  auto solMin = base_nodes->GetSolution_Min();
+  auto solMax = base_nodes->GetSolution_Max();
+  auto limiter = base_nodes->GetLimiter();
 
-      for (iVar = 0; iVar < nVar; iVar++) {
-        
-        AD::SetPreaccIn(base_nodes->GetSolution_Max(iPoint, iVar));
-        AD::SetPreaccIn(base_nodes->GetSolution_Min(iPoint, iVar));
-        AD::SetPreaccIn(base_nodes->GetSolution_Max(jPoint,iVar));
-        AD::SetPreaccIn(base_nodes->GetSolution_Min(jPoint,iVar));
-
-        /*--- Calculate the interface left gradient, delta- (dm) ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_j[iDim]-Coord_i[iDim])*Gradient_i[iVar][iDim];
-        
-        if (dm == 0.0) { limiter = 2.0; }
-        else {
-          if ( dm > 0.0 ) dp = base_nodes->GetSolution_Max(iPoint, iVar);
-          else dp = base_nodes->GetSolution_Min(iPoint, iVar);
-          limiter = dp/dm;
-        }
-        
-        if (limiter < base_nodes->GetLimiter(iPoint, iVar)) {
-          base_nodes->SetLimiter(iPoint, iVar, limiter);
-          AD::SetPreaccOut(base_nodes->GetLimiter(iPoint)[iVar]);
-        }
-        
-        /*--- Calculate the interface right gradient, delta+ (dp) ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_i[iDim]-Coord_j[iDim])*Gradient_j[iVar][iDim];
-        
-        if (dm == 0.0) { limiter = 2.0; }
-        else {
-          if ( dm > 0.0 ) dp = base_nodes->GetSolution_Max(jPoint,iVar);
-          else dp = base_nodes->GetSolution_Min(jPoint,iVar);
-          limiter = dp/dm;
-        }
-        
-        if (limiter < base_nodes->GetLimiter(jPoint,iVar)) {
-          base_nodes->SetLimiter(jPoint,iVar, limiter);
-          AD::SetPreaccOut(base_nodes->GetLimiter(jPoint)[iVar]);
-        }
-
-      }
-      
-      AD::EndPreacc();
-      
-    }
-
-
-    for (iPoint = 0; iPoint < geometry->GetnPoint(); iPoint++) {
-      for (iVar = 0; iVar < nVar; iVar++) {
-        y =  base_nodes->GetLimiter(iPoint, iVar);
-        limiter = (y*y + 2.0*y) / (y*y + y + 2.0);
-        base_nodes->SetLimiter(iPoint, iVar, limiter);
-      }
-    }
-    
-  }
-
-  /*--- Venkatakrishnan limiter ---*/
-  
-  if ((config->GetKind_SlopeLimit() == VENKATAKRISHNAN) || (config->GetKind_SlopeLimit_Flow() == VENKATAKRISHNAN_WANG)) {
-    
-    if (config->GetKind_SlopeLimit_Flow() == VENKATAKRISHNAN_WANG) {
-
-      /*--- Allocate memory for the max and min solution value --*/
-      
-      LocalMinSolution = new su2double [nVar]; GlobalMinSolution = new su2double [nVar];
-      LocalMaxSolution = new su2double [nVar]; GlobalMaxSolution = new su2double [nVar];
-      
-      /*--- Compute the max value and min value of the solution ---*/
-      
-      Solution = base_nodes->GetSolution(iPoint);
-      for (iVar = 0; iVar < nVar; iVar++) {
-        LocalMinSolution[iVar] = Solution[iVar];
-        LocalMaxSolution[iVar] = Solution[iVar];
-      }
-      
-      for (iPoint = 0; iPoint < geometry->GetnPoint(); iPoint++) {
-        
-        /*--- Get the solution variables ---*/
-        
-        Solution = base_nodes->GetSolution(iPoint);
-        
-        for (iVar = 0; iVar < nVar; iVar++) {
-          LocalMinSolution[iVar] = min (LocalMinSolution[iVar], Solution[iVar]);
-          LocalMaxSolution[iVar] = max (LocalMaxSolution[iVar], Solution[iVar]);
-        }
-        
-      }
-      
-#ifdef HAVE_MPI
-      SU2_MPI::Allreduce(LocalMinSolution, GlobalMinSolution, nVar, MPI_DOUBLE, MPI_MIN, MPI_COMM_WORLD);
-      SU2_MPI::Allreduce(LocalMaxSolution, GlobalMaxSolution, nVar, MPI_DOUBLE, MPI_MAX, MPI_COMM_WORLD);
-#else
-      for (iVar = 0; iVar < nVar; iVar++) {
-        GlobalMinSolution[iVar] = LocalMinSolution[iVar];
-        GlobalMaxSolution[iVar] = LocalMaxSolution[iVar];
-      }
-#endif
-    }
-    
-    for (iEdge = 0; iEdge < geometry->GetnEdge(); iEdge++) {
-      
-      iPoint     = geometry->edge[iEdge]->GetNode(0);
-      jPoint     = geometry->edge[iEdge]->GetNode(1);
-      Gradient_i = base_nodes->GetGradient_Reconstruction(iPoint);
-      Gradient_j = base_nodes->GetGradient_Reconstruction(jPoint);
-      Coord_i    = geometry->node[iPoint]->GetCoord();
-      Coord_j    = geometry->node[jPoint]->GetCoord();
-      
-      AD::StartPreacc();
-      AD::SetPreaccIn(Gradient_i, nVar, nDim);
-      AD::SetPreaccIn(Gradient_j, nVar, nDim);
-      AD::SetPreaccIn(Coord_i, nDim); AD::SetPreaccIn(Coord_j, nDim);
-
-      for (iVar = 0; iVar < nVar; iVar++) {
-          
-        AD::StartPreacc();
-        AD::SetPreaccIn(Gradient_i[iVar], nDim);
-        AD::SetPreaccIn(Gradient_j[iVar], nDim);
-        AD::SetPreaccIn(Coord_i, nDim);
-        AD::SetPreaccIn(Coord_j, nDim);
-        AD::SetPreaccIn(base_nodes->GetSolution_Max(iPoint, iVar));
-        AD::SetPreaccIn(base_nodes->GetSolution_Min(iPoint, iVar));
-        AD::SetPreaccIn(base_nodes->GetSolution_Max(jPoint,iVar));
-        AD::SetPreaccIn(base_nodes->GetSolution_Min(jPoint,iVar));
-        
-        if (config->GetKind_SlopeLimit_Flow() == VENKATAKRISHNAN_WANG) {
-          AD::SetPreaccIn(GlobalMaxSolution[iVar]);
-          AD::SetPreaccIn(GlobalMinSolution[iVar]);
-          eps1 = LimK * (GlobalMaxSolution[iVar] - GlobalMinSolution[iVar]);
-          eps2 = eps1*eps1;
-        }
-        else {
-          eps1 = LimK*dave;
-          eps2 = eps1*eps1*eps1;
-        }
-
-        /*--- Calculate the interface left gradient, delta- (dm) ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_j[iDim]-Coord_i[iDim])*Gradient_i[iVar][iDim];
-        
-        /*--- Calculate the interface right gradient, delta+ (dp) ---*/
-        
-        if ( dm > 0.0 ) dp = base_nodes->GetSolution_Max(iPoint, iVar);
-        else dp = base_nodes->GetSolution_Min(iPoint, iVar);
-        
-        limiter = ( dp*dp + 2.0*dp*dm + eps2 )/( dp*dp + dp*dm + 2.0*dm*dm + eps2);
-        
-        if (limiter < base_nodes->GetLimiter(iPoint, iVar)) {
-          base_nodes->SetLimiter(iPoint, iVar, limiter);
-          AD::SetPreaccOut(base_nodes->GetLimiter(iPoint)[iVar]);
-        }
-        
-        /*-- Repeat for point j on the edge ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_i[iDim]-Coord_j[iDim])*Gradient_j[iVar][iDim];
-        
-        if ( dm > 0.0 ) dp = base_nodes->GetSolution_Max(jPoint,iVar);
-        else dp = base_nodes->GetSolution_Min(jPoint,iVar);
-        
-        limiter = ( dp*dp + 2.0*dp*dm + eps2 )/( dp*dp + dp*dm + 2.0*dm*dm + eps2);
-        
-        if (limiter < base_nodes->GetLimiter(jPoint,iVar)) {
-          base_nodes->SetLimiter(jPoint,iVar, limiter);
-          AD::SetPreaccOut(base_nodes->GetLimiter(jPoint)[iVar]);
-        }
-        
-        AD::EndPreacc();
-      }
-    }
-    
-    if (LocalMinSolution  != NULL) delete [] LocalMinSolution;
-    if (LocalMaxSolution  != NULL) delete [] LocalMaxSolution;
-    if (GlobalMinSolution != NULL) delete [] GlobalMinSolution;
-    if (GlobalMaxSolution != NULL) delete [] GlobalMaxSolution;
-
-  }
-  
-  /*--- Sharp edges limiter ---*/
-  
-  if (config->GetKind_SlopeLimit() == SHARP_EDGES) {
-    
-    /*-- Get limiter parameters from the configuration file ---*/
-    
-    dave = config->GetRefElemLength();
-    LimK = config->GetVenkat_LimiterCoeff();
-    eps1 = LimK*dave;
-    eps2 = eps1*eps1*eps1;
-    
-    for (iEdge = 0; iEdge < geometry->GetnEdge(); iEdge++) {
-      
-      iPoint     = geometry->edge[iEdge]->GetNode(0);
-      jPoint     = geometry->edge[iEdge]->GetNode(1);
-      Gradient_i = base_nodes->GetGradient_Reconstruction(iPoint);
-      Gradient_j = base_nodes->GetGradient_Reconstruction(jPoint);
-      Coord_i    = geometry->node[iPoint]->GetCoord();
-      Coord_j    = geometry->node[jPoint]->GetCoord();
-      
-      for (iVar = 0; iVar < nVar; iVar++) {
-        
-        /*--- Calculate the interface left gradient, delta- (dm) ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_j[iDim]-Coord_i[iDim])*Gradient_i[iVar][iDim];
-        
-        /*--- Calculate the interface right gradient, delta+ (dp) ---*/
-        
-        if ( dm > 0.0 ) dp = base_nodes->GetSolution_Max(iPoint, iVar);
-        else dp = base_nodes->GetSolution_Min(iPoint, iVar);
-        
-        /*--- Compute the distance to a sharp edge ---*/
-        
-        SharpEdge_Distance = (geometry->node[iPoint]->GetSharpEdge_Distance() - config->GetAdjSharp_LimiterCoeff()*eps1);
-        ds = 0.0;
-        if (SharpEdge_Distance < -eps1) ds = 0.0;
-        if (fabs(SharpEdge_Distance) <= eps1) ds = 0.5*(1.0+(SharpEdge_Distance/eps1)+(1.0/PI_NUMBER)*sin(PI_NUMBER*SharpEdge_Distance/eps1));
-        if (SharpEdge_Distance > eps1) ds = 1.0;
-        
-        limiter = ds * ( dp*dp + 2.0*dp*dm + eps2 )/( dp*dp + dp*dm + 2.0*dm*dm + eps2);
-        
-        if (limiter < base_nodes->GetLimiter(iPoint, iVar))
-          base_nodes->SetLimiter(iPoint, iVar, limiter);
-        
-        /*-- Repeat for point j on the edge ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_i[iDim]-Coord_j[iDim])*Gradient_j[iVar][iDim];
-        
-        if ( dm > 0.0 ) dp = base_nodes->GetSolution_Max(jPoint,iVar);
-        else dp = base_nodes->GetSolution_Min(jPoint,iVar);
-        
-        /*--- Compute the distance to a sharp edge ---*/
-        
-        SharpEdge_Distance = (geometry->node[jPoint]->GetSharpEdge_Distance() - config->GetAdjSharp_LimiterCoeff()*eps1);
-        ds = 0.0;
-        if (SharpEdge_Distance < -eps1) ds = 0.0;
-        if (fabs(SharpEdge_Distance) <= eps1) ds = 0.5*(1.0+(SharpEdge_Distance/eps1)+(1.0/PI_NUMBER)*sin(PI_NUMBER*SharpEdge_Distance/eps1));
-        if (SharpEdge_Distance > eps1) ds = 1.0;
-        
-        limiter = ds * ( dp*dp + 2.0*dp*dm + eps2 )/( dp*dp + dp*dm + 2.0*dm*dm + eps2);
-        
-        if (limiter < base_nodes->GetLimiter(jPoint,iVar))
-          base_nodes->SetLimiter(jPoint,iVar, limiter);
-        
-      }
-    }
-  }
-  
-  /*--- Sharp edges limiter ---*/
-  
-  if (config->GetKind_SlopeLimit() == WALL_DISTANCE) {
-    
-    /*-- Get limiter parameters from the configuration file ---*/
-    
-    dave = config->GetRefElemLength();
-    LimK = config->GetVenkat_LimiterCoeff();
-    eps1 = LimK*dave;
-    eps2 = eps1*eps1*eps1;
-    
-    for (iEdge = 0; iEdge < geometry->GetnEdge(); iEdge++) {
-      
-      iPoint     = geometry->edge[iEdge]->GetNode(0);
-      jPoint     = geometry->edge[iEdge]->GetNode(1);
-      Gradient_i = base_nodes->GetGradient_Reconstruction(iPoint);
-      Gradient_j = base_nodes->GetGradient_Reconstruction(jPoint);
-      Coord_i    = geometry->node[iPoint]->GetCoord();
-      Coord_j    = geometry->node[jPoint]->GetCoord();
-      
-      for (iVar = 0; iVar < nVar; iVar++) {
-        
-        /*--- Calculate the interface left gradient, delta- (dm) ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_j[iDim]-Coord_i[iDim])*Gradient_i[iVar][iDim];
-        
-        /*--- Calculate the interface right gradient, delta+ (dp) ---*/
-        
-        if ( dm > 0.0 ) dp = base_nodes->GetSolution_Max(iPoint, iVar);
-        else dp = base_nodes->GetSolution_Min(iPoint, iVar);
-        
-        /*--- Compute the distance to a sharp edge ---*/
-        
-        SharpEdge_Distance = (geometry->node[iPoint]->GetWall_Distance() - config->GetAdjSharp_LimiterCoeff()*eps1);
-        ds = 0.0;
-        if (SharpEdge_Distance < -eps1) ds = 0.0;
-        if (fabs(SharpEdge_Distance) <= eps1) ds = 0.5*(1.0+(SharpEdge_Distance/eps1)+(1.0/PI_NUMBER)*sin(PI_NUMBER*SharpEdge_Distance/eps1));
-        if (SharpEdge_Distance > eps1) ds = 1.0;
-        
-        limiter = ds * ( dp*dp + 2.0*dp*dm + eps2 )/( dp*dp + dp*dm + 2.0*dm*dm + eps2);
-        
-        if (limiter < base_nodes->GetLimiter(iPoint, iVar))
-          base_nodes->SetLimiter(iPoint, iVar, limiter);
-        
-        /*-- Repeat for point j on the edge ---*/
-        
-        dm = 0.0;
-        for (iDim = 0; iDim < nDim; iDim++)
-          dm += 0.5*(Coord_i[iDim]-Coord_j[iDim])*Gradient_j[iVar][iDim];
-        
-        if ( dm > 0.0 ) dp = base_nodes->GetSolution_Max(jPoint,iVar);
-        else dp = base_nodes->GetSolution_Min(jPoint,iVar);
-        
-        /*--- Compute the distance to a sharp edge ---*/
-        
-        SharpEdge_Distance = (geometry->node[jPoint]->GetWall_Distance() - config->GetAdjSharp_LimiterCoeff()*eps1);
-        ds = 0.0;
-        if (SharpEdge_Distance < -eps1) ds = 0.0;
-        if (fabs(SharpEdge_Distance) <= eps1) ds = 0.5*(1.0+(SharpEdge_Distance/eps1)+(1.0/PI_NUMBER)*sin(PI_NUMBER*SharpEdge_Distance/eps1));
-        if (SharpEdge_Distance > eps1) ds = 1.0;
-        
-        limiter = ds * ( dp*dp + 2.0*dp*dm + eps2 )/( dp*dp + dp*dm + 2.0*dm*dm + eps2);
-        
-        if (limiter < base_nodes->GetLimiter(jPoint,iVar))
-          base_nodes->SetLimiter(jPoint,iVar, limiter);
-        
-      }
-    }
-  }
-
-  /*--- Correct the limiter values across any periodic boundaries. ---*/
-
-  for (unsigned short iPeriodic = 1; iPeriodic <= config->GetnMarker_Periodic()/2; iPeriodic++) {
-    InitiatePeriodicComms(geometry, config, iPeriodic, PERIODIC_LIM_SOL_2);
-    CompletePeriodicComms(geometry, config, iPeriodic, PERIODIC_LIM_SOL_2);
-  }
-  
-  /*--- Limiter MPI ---*/
-  
-  InitiateComms(geometry, config, SOLUTION_LIMITER);
-  CompleteComms(geometry, config, SOLUTION_LIMITER);
-
-#ifdef CODI_REVERSE_TYPE
-  if (TapeActive) AD::StartRecording();
-#endif
+  computeLimiters(kindLimiter, this, SOLUTION_LIMITER, PERIODIC_LIM_SOL_1, PERIODIC_LIM_SOL_2,
+                  *geometry, *config, 0, nVar, *solution, *gradient, *solMin, *solMax, *limiter);
 }
 
 void CSolver::Gauss_Elimination(su2double** A, su2double* rhs, unsigned short nVar) {
